[8.4.x] Optimization for the Post Drupal Scaffold Procedure.

+++ scripts/varbase/Procedures.php
@@ -59,25 +59,26 @@ class Procedures {
     $root = static::getDrupalRoot(getcwd());
 
     if ($fs->exists($root . '/profiles/varbase/src/assets/robots-staging.txt')) {
-      //Create staging robots file
+      // Create staging robots file.
       copy($root . '/profiles/varbase/src/assets/robots-staging.txt', $root . '/robots-staging.txt');
     }
     
-    if ($fs->exists($root . '/.htaccess')) {
-      //Alter .htaccess file
+    if ($fs->exists($root . '/.htaccess')
+      && $fs->exists($root . '/profiles/varbase/src/assets/htaccess_extra')) {
+
+      // Alter .htaccess file.
       $htaccess_path = $root . '/.htaccess';
       $htaccess_lines = file($htaccess_path);
       $lines = [];
       foreach ($htaccess_lines as $line) {
         $lines[] = $line;
-        if (strpos($line, "RewriteEngine on") !== FALSE
-          && $fs->exists($root . '/profiles/varbase/src/assets/htaccess_extra')) {
+        if (strpos($line, "RewriteEngine on") !== FALSE) {
           $lines = array_merge($lines, file($root . '/profiles/varbase/src/assets/htaccess_extra'));
         }
       }
       file_put_contents($htaccess_path, $lines);
     }
-    
+
     if ($fs->exists($root . '/profiles/varbase/src/assets/development.services.yml')) {
       // Alter development.services.yml to have Varbase's Local development services.
       copy($root . '/profiles/varbase/src/assets/development.services.yml', $root . '/sites/development.services.yml');
